+++ frontend/src/api/index.ts
@@ -1,6 +1,5 @@
 export { API_URL } from './config';
 export { QUERY_KEY } from './queryKeys';
-
 export {
   PAGE_SIZE,
   SORTING_OPTIONS,
@@ -13,7 +12,7 @@ export {
   deleteTemplate,
 } from './templates';
 export { postSignup, postLogin, postLogout, getLoginState, checkName } from './authentication';
-export { getCategoryList, postCategory, deleteCategory } from './categories';
+export { getCategoryList, postCategory, editCategory, deleteCategory } from './categories';
 export { getTagList } from './tags';
 export { postLike, deleteLike } from './like';
 export { getMemberName } from './members';
