+++ frontend/src/api/members.ts
@@ -1,6 +1,5 @@
+import { apiClient } from '@/api/config';
 import { END_POINTS } from '@/routes';
-
-import { apiClient } from './config';
 
 export const getMemberName = async (memberId: number) => {
   const response = await apiClient.get(`${END_POINTS.MEMBERS}/${memberId}/name`);
